+++ gpdb-20160503/alibabacloud_gpdb20160503/client.py
@@ -10399,6 +10399,8 @@ class Client(OpenApiClient):
             query['HybridSearch'] = request.hybrid_search
         if not UtilClient.is_unset(request.hybrid_search_args_shrink):
             query['HybridSearchArgs'] = request.hybrid_search_args_shrink
+        if not UtilClient.is_unset(request.include_metadata_fields):
+            query['IncludeMetadataFields'] = request.include_metadata_fields
         if not UtilClient.is_unset(request.include_values):
             query['IncludeValues'] = request.include_values
         if not UtilClient.is_unset(request.metrics):
@@ -10407,6 +10409,10 @@ class Client(OpenApiClient):
             query['Namespace'] = request.namespace
         if not UtilClient.is_unset(request.namespace_password):
             query['NamespacePassword'] = request.namespace_password
+        if not UtilClient.is_unset(request.offset):
+            query['Offset'] = request.offset
+        if not UtilClient.is_unset(request.order_by):
+            query['OrderBy'] = request.order_by
         if not UtilClient.is_unset(request.owner_id):
             query['OwnerId'] = request.owner_id
         if not UtilClient.is_unset(request.region_id):
@@ -10459,6 +10465,8 @@ class Client(OpenApiClient):
             query['HybridSearch'] = request.hybrid_search
         if not UtilClient.is_unset(request.hybrid_search_args_shrink):
             query['HybridSearchArgs'] = request.hybrid_search_args_shrink
+        if not UtilClient.is_unset(request.include_metadata_fields):
+            query['IncludeMetadataFields'] = request.include_metadata_fields
         if not UtilClient.is_unset(request.include_values):
             query['IncludeValues'] = request.include_values
         if not UtilClient.is_unset(request.metrics):
@@ -10467,6 +10475,10 @@ class Client(OpenApiClient):
             query['Namespace'] = request.namespace
         if not UtilClient.is_unset(request.namespace_password):
             query['NamespacePassword'] = request.namespace_password
+        if not UtilClient.is_unset(request.offset):
+            query['Offset'] = request.offset
+        if not UtilClient.is_unset(request.order_by):
+            query['OrderBy'] = request.order_by
         if not UtilClient.is_unset(request.owner_id):
             query['OwnerId'] = request.owner_id
         if not UtilClient.is_unset(request.region_id):
@@ -10537,6 +10549,8 @@ class Client(OpenApiClient):
             query['HybridSearch'] = request.hybrid_search
         if not UtilClient.is_unset(request.hybrid_search_args_shrink):
             query['HybridSearchArgs'] = request.hybrid_search_args_shrink
+        if not UtilClient.is_unset(request.include_metadata_fields):
+            query['IncludeMetadataFields'] = request.include_metadata_fields
         if not UtilClient.is_unset(request.include_vector):
             query['IncludeVector'] = request.include_vector
         if not UtilClient.is_unset(request.metrics):
@@ -10605,6 +10619,8 @@ class Client(OpenApiClient):
             query['HybridSearch'] = request.hybrid_search
         if not UtilClient.is_unset(request.hybrid_search_args_shrink):
             query['HybridSearchArgs'] = request.hybrid_search_args_shrink
+        if not UtilClient.is_unset(request.include_metadata_fields):
+            query['IncludeMetadataFields'] = request.include_metadata_fields
         if not UtilClient.is_unset(request.include_vector):
             query['IncludeVector'] = request.include_vector
         if not UtilClient.is_unset(request.metrics):